requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
scikit-learn==1.5.2
pandas==2.2.3
joblib==1.4.2
numpy==2.1.1
streamlit==1.39.0
requests==2.32.3
python-dotenv==1.0.1
streamlit
requests
python-dotenv
scikit-learn
pandas
numpy